% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmabasket-package.R
\docType{package}
\name{bmabasket-package}
\alias{bmabasket}
\alias{bmabasket-package}
\title{bmabasket: Bayesian Model Averaging for Basket Trials}
\description{
An implementation of the Bayesian model averaging method 
    of Psioda and others (2019) <doi:10.1093/biostatistics/kxz014> for basket trials. 
    Contains a user-friendly wrapper for simulating basket trials under conditions 
    and analyzing them with a Bayesian model averaging approach.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ethan-alt/bmabasket}
  \item Report bugs at \url{https://github.com/ethan-alt/bmabasket/issues}
}

}
\author{
\strong{Maintainer}: Matt Psioda \email{matt_psioda@unc.edu}

Authors:
\itemize{
  \item Ethan Alt \email{ethanalt@live.unc.edu}
}

}
\keyword{internal}
