% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{numModels}
\alias{numModels}
\title{Compute number of models}
\usage{
numModels(K, P)
}
\arguments{
\item{K}{positive integer giving number of baskets}

\item{P}{positive integer giving maximal number of distinct rates}
}
\value{
integer giving number of possible models
}
\description{
Given a basket size and maximal number of distinct
response rates, compute the number of possible models
}
\examples{
numModels(10, 10)

}
