% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{bma}
\alias{bma}
\title{Compute posterior model probabilities}
\usage{
bma(x, y, n, P = NULL, mu0 = 0.5, phi0 = 1, priorModelProbs = NULL, alpha = 1)
}
\arguments{
\item{x}{scalar between 0 and 1 to evaluate posterior probability}

\item{y}{vector of responses}

\item{n}{vector of sample sizes}

\item{P}{integer giving maximum partition size; default is all partitions}

\item{mu0}{prior mean for beta prior}

\item{phi0}{prior dispersion for beta prior}

\item{priorModelProbs}{vector giving prior for models. Default is prior of each model is proportional to the number of unique probabilities raised to the power alpha}

\item{alpha}{nonnegative scalar. Value of 0 corresponds to uniform prior across model space. Ignored if priorModelProbs is specified.}
}
\description{
Given data and hyperparameters, computes posterior model probabilities
}
\examples{
## Simulate data with 3 baskets
probs <- c(0.5, 0.25, 0.25)
n <- rep(100, length(probs))
y <- rbinom(length(probs), size = n, prob = probs)
bma(0.5, y, n)

}
