% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma_design.R
\name{bma_design}
\alias{bma_design}
\title{Simulate a BMA design}
\usage{
bma_design(
  nSims,
  nBaskets,
  maxDistinct = nBaskets,
  eRates,
  rRates,
  meanTime,
  sdTime,
  ppEffCrit = NULL,
  ppFutCrit,
  futOnly = FALSE,
  rRatesNull,
  rRatesAlt,
  minSSFut,
  minSSEff,
  minSSEnr,
  maxSSEnr,
  targSSPer,
  I0,
  mu0 = 0.5,
  phi0 = 1,
  priorModelProbs = NULL,
  pmp0 = 1
)
}
\arguments{
\item{nSims}{number of simulation studies to be performed}

\item{nBaskets}{number of baskets}

\item{maxDistinct}{integer between 1 and \code{nBaskets} giving number of distinct model probabilities to use. Defaults to \code{nBaskets}. It is recommended to call \code{\link{numModels}} to ensure that computation is tractable.}

\item{eRates}{scalar or vector of Poisson process rates for each basket}

\item{rRates}{scalar or vector of true response rates for each basket}

\item{meanTime}{mean parameter for time to outcome ascertainment}

\item{sdTime}{standard deviation parameter for time to outcome ascertainment}

\item{ppEffCrit}{scalar or vector giving basket-specific posterior probability threshold for activity (i.e., efficacy). Must be specified if \code{futOnly==FALSE}}

\item{ppFutCrit}{scalar or vector giving basket-specific posterior probability threshold for futility}

\item{futOnly}{\code{logical} giving whether design allows only for futility stopping (\code{TRUE} = futility only, \code{FALSE} = both futility and efficacy)}

\item{rRatesNull}{scalar or vector of basket-specific null hypothesis values (for efficacy determination)}

\item{rRatesAlt}{scalar or vector of basket-specific hypothesized alternative values (for futility determination)}

\item{minSSFut}{minimum number of subjects in basket to assess futility}

\item{minSSEff}{minimum number of subjects in basket to assess activity}

\item{minSSEnr}{matrix giving minimum number of new subjects per basket before next analysis (each row is an interim analysis, each column is a basket)}

\item{maxSSEnr}{matrix giving maximum number of new subjects per basket before next analysis (each row is an interim analysis, each column is a basket)}

\item{targSSPer}{scalar or vector giving target sample size increment for each basket}

\item{I0}{maximum number of analyses}

\item{mu0}{prior mean for the response probabilities}

\item{phi0}{prior dispersion response probabilities}

\item{priorModelProbs}{vector giving prior probabilities for models. Default is prior of each model is proportional to the number of unique probabilities raised to the power \code{pmp0}}

\item{pmp0}{scalar giving power for \code{priorModelProbs}. If \code{pmp0==0}, a uniform prior is used for model probabilities. Defaults to 1. Ignored if \code{priorModelProbs} is not \code{NULL}}
}
\value{
a list giving aspects of the simulation
}
\description{
Simulates a BMA design given hyperparameters
}
