% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmabasket-package.R
\docType{package}
\name{bmabasket-package}
\alias{bmabasket}
\alias{bmabasket-package}
\title{bmabasket: Bayesian model averaging for basket trials}
\description{
Implementation of the Bayesian model average method of
             Psioda et al. (2019) <doi:10.1093/biostatistics/kxz014>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ethan-alt/bmabasket}
  \item Report bugs at \url{https://github.com/ethan-alt/bmabasket/issues}
}

}
\author{
\strong{Maintainer}: Ethan Alt \email{ethanalt@live.unc.edu}

Authors:
\itemize{
  \item Matthew Psioda \email{matt_psioda@unc.edu}
}

}
\keyword{internal}
